requirements.txt
# Deep Learning & ML
tensorflow>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Image processing
opencv-python>=4.8.0
pillow>=10.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Cloud storage
cloudinary>=1.34.0

# LLM Integration
groq>=0.4.0

# GUI
customtkinter>=5.2.0

# Utilities
python-dotenv>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
